==> 01_boardgames/data/02_boardgame_list_cleaning.ipynb <==
import pandas as pd

df = pd.read_csv(r'raw\boardgames_list.csv')
df.head()

df.info() # 150057 

link_duplicates = df.loc[df['link'].duplicated(), 'link']
df[df['link'].isin(link_duplicates)].sort_values('link').head(30)

df.drop_duplicates(subset=['link'], keep='last', inplace=True)
len(df)

df[df['rank'].isnull()].sort_values('num_voters', ascending=False).head()

df[df['link'].str.contains('boardgameexpansion')].sort_values('rank').head()

df = df[~df['link'].str.contains('boardgameexpansion')]
len(df)

df[df['rank'].isnull()].sort_values('num_voters', ascending=False).head(30)

df = df[~df['rank'].isnull()]
len(df)

df.info()

df['rank'] = df['rank'].astype(int)
df['year'] = df['year'].astype('Int64')
df['num_voters'] = df['num_voters'].astype(int)

df['id'] = df['link'].str.split('/', expand=True)[4].astype(int)

df[df['price'].notnull() & ~df['price'].fillna('').str.contains('€')]

df[df['price'].notnull() & ~df['price'].fillna('').str.contains('Amazon')]

df['price_euro'] = df['price'].str.split(expand=True)[1].str.replace('.', '', regex=False)\
                              .str.replace(',', '.', regex=False).astype(float)

df[df['price'].notnull() & df['price_euro'].isnull()]

df.columns

columns = ['id', 'title', 'rank', 'link', 'year', 'geek_rating', 'avg_rating', 'num_voters', 'description', 'price_euro']
df = df[columns]

df.head()

sum(df['id'].duplicated())

df.info()  # 25447 

df.describe()

df.sort_values('year').head(10)

df.to_csv(r'raw\boardgames_list_clean.csv', index=False)
